+++ models/todo.js
@@ -1,0 +1,17 @@
+
+const mysql = require('mysql');
+
+
+const db = mysql.createConnection({
+    host : 'localhost',
+    user : 'admin',
+    password : 'avacado1219',
+    database : 'ninjatasker'
+})
+
+db.connect((err) => {
+     if (err) throw err;
+     console.log("db is connected")
+})
+
+module.exports = db;
