+++ routes/index.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const todo = require('./todo');
+
+const router = express.Router();
+
+router.use(todo);
+
+module.exports = router;
